requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
Pillow==10.1.0
numpy==1.26.2
torch==2.1.0
torchvision==0.16.0
timm==0.9.12
